refactor: add character package
